requirements.txt
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
kiteconnect>=4.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
